+++ config/app.php
@@ -2,6 +2,6 @@
 return [
  'name'=>'MR BAR',
  'timezone'=>'Asia/Bangkok',
- 'version'=>'1.48.51',
- 'pack'=>'MR BAR TIME Mobile Recovery v1.48.51',
+ 'version'=>'1.48.52',
+ 'pack'=>'MR BAR TIME HTTPS Landing v1.48.52',
 ];